split to multi modules to reduce dupicate codes between web and app framewok. - core module. - spi module.
