+++ examples/cluster-example/data/src/main/resources/META-INF/services/com.origin.framework.spi.OriginTask
