+++ examples/cluster-example/es-data/src/main/resources/META-INF/services/com.origin.framework.spi.OriginRouter
