+++ examples/cluster-example/redis-data/src/main/java/com/kevin/sample/redis/router/DataRouter.java
@@ -1,10 +1,11 @@
 package com.kevin.sample.redis.router;
 
+import com.origin.framework.core.bean.OriginConfig;
+import com.origin.framework.core.bean.OriginWebVertxContext;
+import com.origin.framework.core.handler.AsyncResultHandler;
+import com.origin.framework.spi.OriginRouter;
 import com.origin.starter.web.OriginWebApplication;
-import com.origin.starter.web.domain.OriginConfig;
-import com.origin.starter.web.domain.OriginVertxContext;
-import com.origin.starter.web.handler.AsyncResultHandler;
-import com.origin.starter.web.spi.OriginRouter;
+
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
 import io.vertx.redis.client.*;
@@ -16,7 +17,7 @@ import java.util.function.Function;
 public class DataRouter implements OriginRouter {
 
     @Override
-    public void router(OriginVertxContext originVertxContext, OriginConfig originConfig) {
+    public void router(OriginWebVertxContext originVertxContext, OriginConfig originConfig) {
         Future<RedisConnection> redisConnectionFuture = OriginWebApplication.getBeanFactory().getRedisClient().connect();
         AsyncResultHandler.handleFuture(redisConnectionFuture, redisConnection -> {
             originConfig.getEventBus().consumer("data")
@@ -42,7 +43,7 @@ public class DataRouter implements OriginRouter {
     }
 
 
-    private void setDataV2(OriginVertxContext originVertxContext, RedisConnection connection, String key, long o, Runnable handleData) {
+    private void setDataV2(OriginWebVertxContext originVertxContext, RedisConnection connection, String key, long o, Runnable handleData) {
 
 
         Request request = Request.cmd(Command.SADD).arg(key).arg(o);
@@ -53,7 +54,7 @@ public class DataRouter implements OriginRouter {
     }
 
 
-    private void getData(OriginVertxContext originVertxContext, String key, Function<Response, Void> handleData) {
+    private void getData(OriginWebVertxContext originVertxContext, String key, Function<Response, Void> handleData) {
         OriginWebApplication.getBeanFactory().getRedisClient()
                 .connect().onComplete(ar -> {
                     if (ar.succeeded()) {
