--- examples/cluster-example/redis-data/src/main/resources/META-INF/services/archetype.xml
@@ -1,9 +1,0 @@
-<archetype>
-  <id>redis-data</id>
-  <sources>
-    <source>src/main/java/App.java</source>
-  </sources>
-  <testSources>
-    <source>src/test/java/AppTest.java</source>
-  </testSources>
-</archetype>
