+++ examples/single-example/src/main/resources/META-INF/services/com.origin.framework.spi.OriginRouter
