+++ origin-framework-core/src/main/java/com/origin/framework/core/exception/InvalidArgumentException.java
@@ -1,0 +1,11 @@
+package com.origin.framework.core.exception;
+
+public class InvalidArgumentException extends RuntimeException{
+    public InvalidArgumentException(String message) {
+        super(message);
+    }
+
+    public InvalidArgumentException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
