+++ origin-framework-core/src/main/java/com/origin/framework/core/exception/InvalidBeanException.java
@@ -1,0 +1,11 @@
+package com.origin.framework.core.exception;
+
+public class InvalidBeanException extends RuntimeException{
+    public InvalidBeanException(String message) {
+        super(message);
+    }
+
+    public InvalidBeanException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
