+++ origin-framework-core/src/main/java/com/origin/framework/core/exception/InvalidResponseException.java
@@ -1,0 +1,11 @@
+package com.origin.framework.core.exception;
+
+public class InvalidResponseException extends RuntimeException{
+    public InvalidResponseException(String message) {
+        super(message);
+    }
+
+    public InvalidResponseException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
