+++ origin-framework-spi/pom.xml
@@ -2,11 +2,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.originframework</groupId>
-    <artifactId>origin-starter-app</artifactId>
+    <artifactId>origin-framework-spi</artifactId>
     <version>2.0</version>
-    <name>origin-starter-app</name>
+    <name>Origin Framework SPI</name>
     <url>https://kevin913.com.cn</url>
-
     <licenses>
         <license>
             <name>Apache License, Version 2.0</name>
@@ -27,27 +26,8 @@
     <dependencies>
         <dependency>
             <groupId>com.originframework</groupId>
-            <artifactId>origin-framework-spi</artifactId>
+            <artifactId>origin-framework-core</artifactId>
             <version>${origin.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <scope>compile</scope>
-            <version>2.15.0</version>
-        </dependency>
-        <dependency>
-            <groupId>io.vertx</groupId>
-            <artifactId>vertx-zookeeper</artifactId>
-            <version>${vertx.version}</version>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <optional>true</optional>
-            <version>1.18.22</version>
         </dependency>
     </dependencies>
     <build>
