+++ origin-framework-spi/src/main/java/com/origin/framework/spi/OriginTask.java
@@ -1,0 +1,9 @@
+package com.origin.framework.spi;
+
+
+import com.origin.framework.core.bean.OriginVertxContext;
+import com.origin.framework.core.bean.OriginConfig;
+
+public interface OriginTask {
+    public void run(OriginVertxContext vc, OriginConfig cf);
+}
