+++ origin-starter-app/src/main/java/com/origin/starter/app/core/OriginTaskFactory.java
@@ -1,9 +1,10 @@
 package com.origin.starter.app.core;
 
 
-import com.origin.starter.app.domain.OriginAppConfig;
-import com.origin.starter.app.domain.OriginAppVertxContext;
-import com.origin.starter.app.spi.OriginTask;
+
+import com.origin.framework.core.bean.OriginVertxContext;
+import com.origin.framework.core.bean.OriginConfig;
+import com.origin.framework.spi.OriginTask;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,11 +13,11 @@ import java.util.ServiceLoader;
 public class OriginTaskFactory {
 
     private final List<OriginTask> routers = new ArrayList<>();
-    private final OriginAppVertxContext vc;
-    private final OriginAppConfig cf;
+    private final OriginVertxContext vc;
+    private final OriginConfig cf;
 
 
-    public OriginTaskFactory(OriginAppVertxContext vc, OriginAppConfig cf) {
+    public OriginTaskFactory(OriginVertxContext vc, OriginConfig cf) {
         this.vc = vc;
         this.cf = cf;
         ServiceLoader<OriginTask> loader = ServiceLoader.load(OriginTask.class);
