--- origin-starter-web/src/main/java/com/origin/starter/web/exception/InvalidArgumentException.java
@@ -1,11 +1,0 @@
-package com.origin.starter.web.exception;
-
-public class InvalidArgumentException extends RuntimeException{
-    public InvalidArgumentException(String message) {
-        super(message);
-    }
-
-    public InvalidArgumentException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
