--- origin-starter-web/src/main/java/com/origin/starter/web/exception/InvalidBeanException.java
@@ -1,11 +1,0 @@
-package com.origin.starter.web.exception;
-
-public class InvalidBeanException extends RuntimeException{
-    public InvalidBeanException(String message) {
-        super(message);
-    }
-
-    public InvalidBeanException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
