--- origin-starter-web/src/main/java/com/origin/starter/web/exception/InvalidResponseException.java
@@ -1,11 +1,0 @@
-package com.origin.starter.web.exception;
-
-public class InvalidResponseException extends RuntimeException{
-    public InvalidResponseException(String message) {
-        super(message);
-    }
-
-    public InvalidResponseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
